Enum ruma_events::SyncMessageLikeEvent

pub enum SyncMessageLikeEvent<C: MessageLikeEventContent + RedactContent>
where
    C::Redacted: RedactedMessageLikeEventContent,
{
    Original(OriginalSyncMessageLikeEvent<C>),
    Redacted(RedactedSyncMessageLikeEvent<C::Redacted>),
}

A possibly-redacted message-like event without a room_id.

SyncMessageLikeEvent implements the comparison traits using only the event_id field, a sorted list would be sorted lexicographically based on the event’s EventId.

Variants

Original(OriginalSyncMessageLikeEvent<C>)

Original, unredacted form of the event.

Redacted(RedactedSyncMessageLikeEvent<C::Redacted>)

Redacted form of the event with minimal fields.

Implementations

impl<C> SyncMessageLikeEvent<C>
where C: MessageLikeEventContent + RedactContent, C::Redacted: RedactedMessageLikeEventContent,

pub fn event_type(&self) -> MessageLikeEventType

Returns the type of this event.

pub fn event_id(&self) -> &EventId

Returns this event’s event_id field.

pub fn sender(&self) -> &UserId

Returns this event’s sender field.

pub fn origin_server_ts(&self) -> MilliSecondsSinceUnixEpoch

Returns this event’s origin_server_ts field.

pub fn as_original(&self) -> Option<&OriginalSyncMessageLikeEvent<C>>

Get the inner OriginalSyncMessageLikeEvent if this is an unredacted event.

pub fn into_full_event(self, room_id: OwnedRoomId) -> MessageLikeEvent<C>

Convert this sync event into a full event (one with a room_id field).