Struct ruma_events::key::verification::accept::SasV1ContentInit

pub struct SasV1ContentInit {
    pub key_agreement_protocol: KeyAgreementProtocol,
    pub hash: HashAlgorithm,
    pub message_authentication_code: MessageAuthenticationCode,
    pub short_authentication_string: Vec<ShortAuthenticationString>,
    pub commitment: Base64,
}

Mandatory initial set of fields for creating an accept SasV1Content.

Fields
key_agreement_protocol: KeyAgreementProtocol

The key agreement protocol the device is choosing to use, out of the options in the m.key.verification.start message.

hash: HashAlgorithm

The hash method the device is choosing to use, out of the options in the m.key.verification.start message.

message_authentication_code: MessageAuthenticationCode

The message authentication codes that the accepting device understands.

short_authentication_string: Vec<ShortAuthenticationString>

The SAS methods both devices involved in the verification process understand.

Must be a subset of the options in the m.key.verification.start message.

commitment: Base64

The hash (encoded as unpadded base64) of the concatenation of the device’s ephemeral public key (encoded as unpadded base64) and the canonical JSON representation of the m.key.verification.start message.