Struct ruma_events::presence::PresenceEventContent

pub struct PresenceEventContent {
    pub avatar_url: Option<String>,
    pub currently_active: bool,
    pub displayname: Option<String>,
    pub last_active_ago: Option<u64>,
    pub presence: PresenceState,
    pub user_id: String,
}

The payload of a PresenceEvent.

Fields
avatar_url: Option<String>

The current avatar URL for this user.

currently_active: bool

Whether or not the user is currently active.

displayname: Option<String>

The current display name for this user.

last_active_ago: Option<u64>

The last time since this used performed some action, in milliseconds.

presence: PresenceState

The presence state for this user.

user_id: String

The unique identifier for the user associated with this event.