Module ruma_events::room::history_visibility

Types for the m.room.history_visibility event.

Structs

HistoryVisibilityEventContent
The payload of a HistoryVisibilityEvent.

Enums

HistoryVisibility
Who can see a room's history.

Type Definitions

HistoryVisibilityEvent
This event controls whether a member of a room can see the events that happened in a room from before they joined.