Enum ruma_events::room::history_visibility::HistoryVisibility

pub enum HistoryVisibility {
    Invited,
    Joined,
    Shared,
    WorldReadable,
}

Who can see a room's history.

Variants
Invited

Previous events are accessible to newly joined members from the point they were invited onwards. Events stop being accessible when the member's state changes to something other than invite or join.

Joined

Previous events are accessible to newly joined members from the point they joined the room onwards. Events stop being accessible when the member's state changes to something other than join.

Shared

Previous events are always accessible to newly joined members. All events in the room are accessible, even those sent when the member was not a part of the room.

WorldReadable

All events while this is the HistoryVisibility value may be shared by any participating homeserver with anyone, regardless of whether they have ever joined the room.