Enum ruma_events::room::join_rules::JoinRule

pub enum JoinRule {
    Invite,
    Knock,
    Private,
    Public,
}

The rule used for users wishing to join this room.

Variants
Invite

A user who wishes to join the room must first receive an invite to the room from someone already inside of the room.

Knock

Reserved but not yet implemented by the Matrix specification.

Private

Reserved but not yet implemented by the Matrix specification.

Public

Anyone can join the room without any prior action.