Struct ruma_events::room::member::RoomMemberEventContent

#[non_exhaustive]
pub struct RoomMemberEventContent {
    pub avatar_url: Option<OwnedMxcUri>,
    pub displayname: Option<String>,
    pub is_direct: Option<bool>,
    pub membership: MembershipState,
    pub third_party_invite: Option<ThirdPartyInvite>,
    pub blurhash: Option<String>,
    pub reason: Option<String>,
    pub join_authorized_via_users_server: Option<OwnedUserId>,
}

The content of an m.room.member event.

The current membership state of a user in the room.

Adjusts the membership state for a user in a room. It is preferable to use the membership APIs (/rooms/<room id>/invite etc) when performing membership actions rather than adjusting the state directly as there are a restricted set of valid transformations. For example, user A cannot force user B to join a room, and trying to force this state change directly will fail.

This event may also include an invite_room_state key inside the event’s unsigned data, but Ruma doesn’t currently expose this; see #998.

The user for which a membership applies is represented by the state_key. Under some conditions, the sender and state_key may not match - this may be interpreted as the sender affecting the membership state of the state_key user.

The membership for a given user can change over time. Previous membership can be retrieved from the prev_content object on an event. If not present, the user’s previous membership must be assumed as leave.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

avatar_url: Option<OwnedMxcUri>

The avatar URL for this user, if any.

This is added by the homeserver. If you activate the compat-empty-string-null feature, this field being an empty string in JSON will result in None here during deserialization.

displayname: Option<String>

The display name for this user, if any.

This is added by the homeserver.

is_direct: Option<bool>

Flag indicating whether the room containing this event was created with the intention of being a direct chat.

membership: MembershipState

The membership state of this user.

third_party_invite: Option<ThirdPartyInvite>

If this member event is the successor to a third party invitation, this field will contain information about that invitation.

blurhash: Option<String>

The BlurHash for the avatar pointed to by avatar_url.

This uses the unstable prefix in MSC2448.

reason: Option<String>

User-supplied text for why their membership has changed.

For kicks and bans, this is typically the reason for the kick or ban. For other membership changes, this is a way for the user to communicate their intent without having to send a message to the room, such as in a case where Bob rejects an invite from Alice about an upcoming concert, but can’t make it that day.

Clients are not recommended to show this reason to users when receiving an invite due to the potential for spam and abuse. Hiding the reason behind a button or other component is recommended.

join_authorized_via_users_server: Option<OwnedUserId>

Arbitrarily chosen UserId (MxID) of a local user who can send an invite.

Implementations

impl RoomMemberEventContent

pub fn new(membership: MembershipState) -> Self

Creates a new RoomMemberEventContent with the given membership state.

pub fn details(&self) -> MembershipDetails<'_>

Obtain the details about this event that are required to calculate a membership change.

This is required when you want to calculate the change a redacted m.room.member event made.

pub fn membership_change<'a>( &'a self, prev_details: Option<MembershipDetails<'a>>, sender: &UserId, state_key: &UserId ) -> MembershipChange<'a>

Helper function for membership change.

This requires data from the full event:

The previous details computed from event.unsigned.prev_content,

The sender of the event,

The state key of the event.

Check the specification for details.