Type Definition ruma_events::room::member::StrippedRoomMemberEvent

pub type StrippedRoomMemberEvent = StrippedStateEvent<PossiblyRedactedRoomMemberEventContent>;

An m.room.member event from an invited room preview.

Implementations

impl StrippedRoomMemberEvent

pub fn details(&self) -> MembershipDetails<'_>

Obtain the details about this event that are required to calculate a membership change.

This is required when you want to calculate the change a redacted m.room.member event made.

pub fn membership_change<'a>( &'a self, prev_details: Option<MembershipDetails<'a>> ) -> MembershipChange<'a>

Helper function for membership change.

Since stripped events don’t have unsigned.prev_content, you have to pass the .details() of the previous m.room.member event manually (if there is a previous m.room.member event).

Check the specification for details.