Type Alias ruma_events::room::power_levels::PossiblyRedactedRoomPowerLevelsEventContent

pub type PossiblyRedactedRoomPowerLevelsEventContent = RoomPowerLevelsEventContent;

The possibly redacted form of RoomPowerLevelsEventContent.

This type is used when it’s not obvious whether the content is redacted or not.

Aliased Type

struct PossiblyRedactedRoomPowerLevelsEventContent {
    pub ban: Int,
    pub events: BTreeMap<TimelineEventType, Int>,
    pub events_default: Int,
    pub invite: Int,
    pub kick: Int,
    pub redact: Int,
    pub state_default: Int,
    pub users: BTreeMap<OwnedUserId, Int>,
    pub users_default: Int,
    pub notifications: NotificationPowerLevels,
}

Fields
ban: Int

The level required to ban a user.

events: BTreeMap<TimelineEventType, Int>

The level required to send specific event types.

This is a mapping from event type to power level required.

events_default: Int

The default level required to send message events.

invite: Int

The level required to invite a user.

kick: Int

The level required to kick a user.

redact: Int

The level required to redact an event.

state_default: Int

The default level required to send state events.

users: BTreeMap<OwnedUserId, Int>

The power levels for specific users.

This is a mapping from user_id to power level for that user.

users_default: Int

The default power level for every user in the room.

notifications: NotificationPowerLevels

The power level requirements for specific notification types.

This is a mapping from key to power level for that notifications key.