Struct ruma_events::room::third_party_invite::ThirdPartyInviteEventContent

pub struct ThirdPartyInviteEventContent {
    pub display_name: String,
    pub key_validity_url: String,
    pub public_key: String,
    pub public_keys: Option<Vec<PublicKey>>,
}

The payload of a ThirdPartyInviteEvent.

Fields
display_name: String

A user-readable string which represents the user who has been invited.

key_validity_url: String

A URL which can be fetched to validate whether the key has been revoked.

public_key: String

A Base64-encoded Ed25519 key with which the token must be signed.

public_keys: Option<Vec<PublicKey>>

Keys with which the token may be signed.