Type Alias ruma_events::space::parent::RedactedSyncSpaceParentEvent

pub type RedactedSyncSpaceParentEvent = RedactedSyncStateEvent<RedactedSpaceParentEventContent>;

An m.space.parent event from a sync_events response that has been redacted.

Aliased Type

struct RedactedSyncSpaceParentEvent {
    pub content: RedactedSpaceParentEventContent,
    pub event_id: OwnedEventId,
    pub sender: OwnedUserId,
    pub origin_server_ts: MilliSecondsSinceUnixEpoch,
    pub state_key: OwnedRoomId,
    pub unsigned: RedactedUnsigned,
}

Fields
content: RedactedSpaceParentEventContent

Data specific to the event type.

event_id: OwnedEventId

The globally unique event identifier for the user who sent the event.

sender: OwnedUserId

The fully-qualified ID of the user who sent this event.

origin_server_ts: MilliSecondsSinceUnixEpoch

Timestamp in milliseconds on originating homeserver when this event was sent.

state_key: OwnedRoomId

A unique key which defines the overwriting semantics for this piece of room state.

This is often an empty string, but some events send a UserId to show which user the event affects.

unsigned: RedactedUnsigned

Additional key-value pairs not signed by the homeserver.