Struct ruma_events::RedactedSyncMessageEvent

pub struct RedactedSyncMessageEvent<C: RedactedMessageEventContent> {
    pub content: C,
    pub event_id: Box<EventId>,
    pub sender: Box<UserId>,
    pub origin_server_ts: MilliSecondsSinceUnixEpoch,
    pub unsigned: RedactedUnsigned,
}

A redacted message event without a room_id.

RedactedSyncMessageEvent implements the comparison traits using only the event_id field, a sorted list would be sorted lexicographically based on the event’s EventId.

Fields
content: C

Data specific to the event type.

event_id: Box<EventId>

The globally unique event identifier for the user who sent the event.

sender: Box<UserId>

The fully-qualified ID of the user who sent this event.

origin_server_ts: MilliSecondsSinceUnixEpoch

Timestamp in milliseconds on originating homeserver when this event was sent.

unsigned: RedactedUnsigned

Additional key-value pairs not signed by the homeserver.

Implementations

impl<C: RedactedMessageEventContent> RedactedSyncMessageEvent<C>

pub fn into_full_event(self, room_id: Box<RoomId>) -> RedactedMessageEvent<C>

Convert this sync event into a full event, one with a room_id field.