Trait ruma_events::EventContent

pub trait EventContent: Sized + Serialize {
    fn event_type(&self) -> &str;

    fn from_parts(event_type: &str, content: &RawJsonValue) -> Result<Self>;
}

The base trait that all event content types implement.

Implementing this trait allows content types to be serialized as well as deserialized.

Required methods

fn event_type(&self) -> &str

A matrix event identifier, like m.room.message.

fn from_parts(event_type: &str, content: &RawJsonValue) -> Result<Self>

Constructs the given event content.