Module ruma_events::room::join_rules

Types for the m.room.join_rules event.

Structs

JoinRulesEventContent
The payload of a JoinRulesEvent.

Enums

JoinRule
The rule used for users wishing to join this room.

Type Definitions

JoinRulesEvent
Describes how users are allowed to join the room.